# Add depth calculation to nodes to make size() constant.