Small thing, but I added a depth property to each node in `Tree.js`, which makes determining the size of the tree operate at constant time. With this efficiency, the Tree `reduce` function can lean on `size()` to be quite a bit faster as well.

These benchmarks are approximate, and don't account for variance, but:

### Before:

```
Nodes  ::append()  ::root()  ::reduce(merge)  ::size()  ::prune()  Memory Growth
-----  ----------  --------  ---------------  --------  ---------  -------------
10000  0.0007ms    0.0240ms  2.84ms           0.55ms    1.13ms     0.476%
```

### After:

```
Nodes  ::append()  ::root()  ::reduce(merge)  ::size()  ::prune()  Memory Growth
-----  ----------  --------  ---------------  --------  ---------  -------------
10000  0.0007ms    0.0210ms  1.81ms           0.03ms    0.85ms     0.447%
```